vktrace: fix warnings introduce by last commit
Also remove unneeded type casts in kreplay_objmapper_class_defs.h, add some comments.

+++ vktrace/vktrace_layer/vktrace_lib_pageguardmappedmemory.cpp
@@ -409,8 +409,8 @@ size_t PageGuardMappedMemory::getChangedBlockInfo(VkDeviceSize RangeOffset, VkDe
         offset = getMappedBlockOffset(i);
         if (isMappedBlockChanged(i, useWhich)) {
             if (pChangedInfoArray) {
-                pChangedInfoArray[dwIndex + 1].offset = offset;
-                pChangedInfoArray[dwIndex + 1].length = CurrentBlockSize;
+                pChangedInfoArray[dwIndex + 1].offset = (uint32_t)offset;
+                pChangedInfoArray[dwIndex + 1].length = (uint32_t)CurrentBlockSize;
                 pChangedInfoArray[dwIndex + 1].reserve0 = 0;
                 pChangedInfoArray[dwIndex + 1].reserve1 = 0;
                 pChangedData = pData + DataOffset + infosize + SaveSize;
@@ -446,8 +446,8 @@ size_t PageGuardMappedMemory::getChangedBlockInfo(VkDeviceSize RangeOffset, VkDe
         }
     }
     if (pChangedInfoArray) {
-        pChangedInfoArray[0].offset = dwAmount;
-        pChangedInfoArray[0].length = SaveSize;
+        pChangedInfoArray[0].offset = (uint32_t)dwAmount;
+        pChangedInfoArray[0].length = (uint32_t)SaveSize;
     }
     if (pdwSaveSize) {
         *pdwSaveSize = SaveSize;
